Fix compiler warning
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@1042 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/memory/memory_pool.cpp
@@ -292,7 +292,7 @@ private:
 	// to keep track of block types
 	typedef std::map<int, int> pool_map_t;
 	pool_map_t*	blk_type_count;
-	int extend_increment;
+	size_t extend_increment;
 };
